Djagnog/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
